% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelGaussian.R
\name{kernelGaussian}
\alias{kernelGaussian}
\title{kernelGaussian: Guassian Kernel used on Promethee IV Kernel.}
\description{
kernelGaussian is a Guassian Kernel used on Promethee IV Kernel.
}
\usage{
kernelGaussian(u)
}
\arguments{
\item{u}{The argument used in kernelGaussian function.}
}
