% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrometheeS4.R
\docType{methods}
\name{NetworkPlot}
\alias{NetworkPlot}
\alias{NetworkPlot,RPrometheeI-method}
\title{NetworkPlot}
\usage{
NetworkPlot(RPrometheeI)
}
\arguments{
\item{RPrometheeI}{An object resulting from RPrometheeI method.}
}
\description{
Shows the relationship among alternatives using a net graph, where the
  arrows come from the alternative with biggest PhiPlus and smallest PhiMinus.
}
\references{
\itemize{

      \item
      J. P. Brans, Ph. Vincke\cr
      \emph{A Preference Ranking Organisation Method: (The PROMETHEE Method
      for Multiple Criteria Decision-Making)}\cr
      Management science, v. 31, n. 6, p. 647-656, 1985.\cr
      \url{https://pdfs.semanticscholar.org/edd6/f5ae9c1bfb2fdd5c9a5d66e56bdb22770460.pdf}

      \item
      J. P. Brans, B. Mareschal \cr
      \emph{PROMETHEE methods. In: Figueria J, Greco S, Ehrgott M (eds)
      Multiple criteria decision analysis: state of the art surveys.}\cr
      Springer Science, Business Media Inc., Boston pp 163???195.\cr
      \url{http://www.springer.com/la/book/9780387230818}
   }
}
\seealso{
Other RPromethee methods: \code{\link{PrometheeIIIPlot}},
  \code{\link{PrometheeIIPlot}},
  \code{\link{PrometheeIPlot}},
  \code{\link{PrometheeIVPlot}},
  \code{\link{RPrometheeConstructor}},
  \code{\link{RPrometheeIII}}, \code{\link{RPrometheeII}},
  \code{\link{RPrometheeIVKernel}},
  \code{\link{RPrometheeIV}}, \code{\link{RPrometheeI}},
  \code{\link{RPrometheeV}},
  \code{\link{SensitivityAnalysis}},
  \code{\link{UpdateRPrometheeAlternatives}},
  \code{\link{UpdateRPrometheeArguments}},
  \code{\link{WalkingWeightsPlot}}
}
\author{
Pedro Henrique Melo Albuquerque, \email{pedroa@unb.br}

Gustavo Monteiro Pereira, \email{monteirogustavop@gmail.com}
}
\keyword{decision-analysis}
\keyword{decision-method}
\keyword{mcda}
\keyword{promethee}
