% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrometheeS4.R
\docType{methods}
\name{plot,RPrometheeI-method}
\alias{plot,RPrometheeI-method}
\title{Plots RPrometheeI objects}
\usage{
\S4method{plot}{RPrometheeI}(x, y, ...)
}
\arguments{
\item{x}{the RPromethee object to be ploted.}

\item{y}{not used in this context.}

\item{...}{not used in this context.}
}
\description{
Plots PhiPlus and PhiMinus resulting from RPrometheeI results.
}
