% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrometheeS4.R
\docType{methods}
\name{plot,RPrometheeII-method}
\alias{plot,RPrometheeII-method}
\title{Plots RPrometheeII objects}
\usage{
\S4method{plot}{RPrometheeII}(x, y, ...)
}
\arguments{
\item{x}{the RPromethee object to be ploted.}

\item{y}{not used in this context.}

\item{...}{not used in this context.}
}
\description{
Plots the net Phi, resulting from RPrometheeII method.
}
