% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrometheeS4.R
\docType{methods}
\name{print,RPrometheeII-method}
\alias{print,RPrometheeII-method}
\title{Prints a RPromethee object.}
\usage{
\S4method{print}{RPrometheeII}(x)
}
\arguments{
\item{x}{A RPromethee object.}

\item{...}{Not used in this context.}
}
\description{
Prints main information from a \code{RPrometheeII} object.
}
