% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PrometheeI}
\alias{PrometheeI}
\title{Create the Kernel matrix}
\usage{
PrometheeI(datMat, vecWeights, prefFunction, parms, normalize)
}
\arguments{
\item{datVec}{Column of the dataset}

\item{int}{Type od preference function}
}
\value{
Preference Matrix
}
\description{
Create the Kernel matrix
}
