% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PrometheeIII}
\alias{PrometheeIII}
\title{Create the Kernel matrix}
\usage{
PrometheeIII(datMat, vecWeights, prefFunction, alphaVector, parms)
}
\arguments{
\item{datVec}{Column of the dataset}

\item{int}{Type od preference function}
}
\value{
Preference Matrix
}
\description{
Create the Kernel matrix
}
