% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMCriteria.R
\docType{package}
\name{RMCriteria-package}
\alias{RMCriteria}
\alias{RMCriteria-package}
\title{Solves MCDA problems using Promethee methods.}
\description{
RMCriteria solves Multiple-Criteria Decision Analysis (MCDA) problems.
For now, it only supports Promethee methods, but other methods may be
developed in the future.
}
\details{
RMCriteria supports Promethees I, II, III, IV, IV Kernel and V.

RMCriteria was developed in the Laboratory of Machine Learning in Finance and
Organizations (LAMFO) from University of Brasilia, in Brazil. LAMFO is a
center devoted to research machine learning methods and related subjects
applied to organizations in Marketing, Finance, Logistics and many others.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/lamfo-unb/RMCriteria}
}

}
