% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMCriteria.R
\docType{package}
\name{RMCriteria}
\alias{RMCriteria}
\alias{RMCriteria-package}
\title{RMCriteria: Solves MCDA problems using Promethee methods.}
\description{
RMCriteria solves Multiple-Criteria Decision Analysis (MCDA) problems.
For now, it only supports Promethee methods, but other methods may be
developed in the future.
}
\details{
RMCriteria supports Promethees I, II, III, IV, IV Kernel and V.
}
