% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrometheeS4.R
\docType{class}
\name{RPrometheeII-class}
\alias{RPrometheeII-class}
\title{An S4 class to store results from RPrometheeII.}
\description{
An S4 class to store results from RPrometheeII.
}
\section{Slots}{

\describe{
\item{\code{Phi}}{A numeric vector with the net Phi from Promethee.}

\item{\code{vecWeights}}{A numeric vector with the weights for each criteria.}

\item{\code{alternatives}}{A character vector with alternatives names.}

\item{\code{criterias}}{A character vector with criterias names.}

\item{\code{datMat}}{A matrix containing the data from criterias and alternatives.}
}}

