% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrometheeS4.R
\name{RPrometheeII}
\alias{RPrometheeII}
\alias{RPrometheeII,RPrometheeArguments-method}
\title{RPrometheeII}
\usage{
RPrometheeII(RPrometheeArguments)
}
\arguments{
\item{RPrometheeArguments}{An object with all RPromethee arguments. See
\code{\link{RPrometheeConstructor}} for more information.}
}
\value{
\itemize{
  \item{Phi} {The resulting net Phi from the alternatives for all
  criterias.}
  \item{alternatives} {The alternatives names.}
  \item{criterias} {The criterias names.}
  \item{datMat} {The data used corresponding to criterias and alternatives.}
 }
}
\description{
Proposed by Brans and Vincke (1985), PROMETHEE II method aims to solve
  sorting problems. The PROMETHEE II method performs a total ordering of the
  alternatives set by calculating the net outranking flow (HENDRIKS et al.,
  1992), with the objective of solving the problem that no unambiguous
  solution can be given due to incomparability.
}
\details{
The method created by Brans et al. (1985) is based on a set of alternatives
  \eqn{A = {a1,a2,...,an}} that will be ordered and a set of criteria
  \eqn{F = { f1, f2, . . ., fm }}. Two alternatives, \eqn{ai} and \eqn{a_j},
  will be pairwise compared. The intensity of the preference between \eqn{ai}
  over \eqn{aj} \eqn{(Pk(dk)}, \eqn{dk = fk (ai) ??? fk (aj))} is determined.
  \eqn{Pk} is considered the preference function for the \eqn{kth} criterion. The evaluation of the alternative \eqn{ai}, which corresponds to criterion
  \eqn{fk}, is \eqn{fk(ai)} (Hsu, Lin, 2014).\cr
  Six types of preference functions were proposed by Brans et al. (1985). The
  preference scales values range from 0 (no preference) to 1 (strong
  preference).\cr
  While anylising the entering and leaving flows, it can be observed that an
  alternative is better than the other when it has the higher leaving flow
  and the lower entering flow. PROMETHEE I method create a partial pre-order
  that can be acquired by comparing the leaving and entering flow (Brans and
  Mareschal 2005).
}
\examples{
## Create objects for each argument
data <-matrix(c(5.2, -3.5,
                4.3, -1.2,
                6.7, -2.0), byrow = TRUE, ncol = 2, nrow = 3)

parms <- matrix(c(NA, NA), byrow = TRUE, ncol = 1, nrow = 2)
vecWeights <- c(0.3, 0.7)
vecMaximiz <- c(FALSE, TRUE)
prefFunction <- c(0, 0)
normalize <- FALSE
alternatives <- c("Alt 1", "Alt 2", "Alt 3")

## Create RPrometheeArguments object
PromObj <- RPrometheeConstructor(datMat = data, vecWeights = vecWeights,
vecMaximiz = vecMaximiz, prefFunction = prefFunction, parms = parms,
normalize = normalize, alternatives = alternatives)

## Run RPrometheeII
(result <- RPrometheeII(PromObj))

## There are two alternatives two plot a RPrometheeII object:
plot(result)
PrometheeIIPlot(result)

## Updating alternatives name using UpdateRPrometheeAlternatives
newAlternatives <- c("A", "B", "C")
result <- UpdateRPrometheeAlternatives(results, newAlternatives)

## Updating any argument using UpdateRPrometheeArguments
newWeights <- c(0.5, 0.5)
result <- UpdateRPrometheeArguments(result, "vecWeights", newWeights)
}
\references{
\itemize{
      \item
      J. P. Brans, Ph. Vincke\cr
      \emph{A Preference Ranking Organisation Method: (The PROMETHEE Method
      for Multiple Criteria Decision-Making)}\cr
      Management science, v. 31, n. 6, p. 647-656, 1985.\cr
      \url{https://pdfs.semanticscholar.org/edd6/f5ae9c1bfb2fdd5c9a5d66e56bdb22770460.pdf}

      \item
      J. P. Brans, B. Mareschal \cr
      \emph{PROMETHEE methods. In: Figueria J, Greco S, Ehrgott M (eds)
      Multiple criteria decision analysis: state of the art surveys.}\cr
      Springer Science, Business Media Inc., Boston pp 163???195.\cr
      \url{http://www.springer.com/la/book/9780387230818}

      \item
      Tsuen-Ho Hsu, Ling-Zhong Lin\cr
      \emph{Using Fuzzy Preference Method for Group Package Tour Based on the
      Risk Perception}.\cr
      Group Decision and Negotiation, v. 23, n. 2, p. 299-323, 2014.\cr
      \url{http://link.springer.com/article/10.1007/s10726-012-9313-7}
   }
}
\seealso{
Other RPromethee methods: \code{\link{NetworkPlot}},
  \code{\link{PrometheeIIIPlot}},
  \code{\link{PrometheeIIPlot}},
  \code{\link{PrometheeIPlot}},
  \code{\link{PrometheeIVPlot}},
  \code{\link{RPrometheeConstructor}},
  \code{\link{RPrometheeIII}},
  \code{\link{RPrometheeIVKernel}},
  \code{\link{RPrometheeIV}}, \code{\link{RPrometheeI}},
  \code{\link{RPrometheeV}},
  \code{\link{SensitivityAnalysis}},
  \code{\link{UpdateRPrometheeAlternatives}},
  \code{\link{UpdateRPrometheeArguments}},
  \code{\link{WalkingWeightsPlot}}
}
\author{
Pedro Henrique Melo Albuquerque, \email{pedroa@unb.br}

Gustavo Monteiro Pereira, \email{monteirogustavop@gmail.com}
}
\keyword{decision-analysis}
\keyword{decision-method}
\keyword{mcda}
\keyword{promethee}
