% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrometheeS4.R
\docType{class}
\name{RPrometheeIII-class}
\alias{RPrometheeIII-class}
\title{An S4 class to store results from RPrometheeIII.}
\description{
An S4 class to store results from RPrometheeIII.
}
\section{Slots}{

\describe{
\item{\code{limInf}}{A numeric vector with the inferior limit for the interval
defined for each flow.}

\item{\code{limSup}}{A numeric vector with the superior limit for the interval
defined for each flow}

\item{\code{Phi}}{A numeric vector with the net Phi from Promethee.}

\item{\code{alternatives}}{A character vector with alternatives names.}

\item{\code{criterias}}{A character vector with criterias names.}

\item{\code{datMat}}{A matrix containing the data from criterias and alternatives.}
}}

