% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrometheeS4.R
\docType{class}
\name{RPrometheeIV-class}
\alias{RPrometheeIV-class}
\title{An S4 class to store results from RPrometheeIV.}
\description{
An S4 class to store results from RPrometheeIV.
}
\section{Slots}{

\describe{
\item{\code{PhiPlus}}{A numeric vector with the PhiPlus result from Promethee.}

\item{\code{PhiMinus}}{A numeric vector with the PhiMinus result from Promethee.}

\item{\code{Index}}{The index resulting from the lp solution.}

\item{\code{alternatives}}{A character vector with alternatives names.}

\item{\code{criterias}}{A character vector with criterias names.}

\item{\code{datMat}}{A matrix containing the data from criterias and alternatives.}
}}

