% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrometheeS4.R
\name{RPrometheeIV}
\alias{RPrometheeIV}
\alias{PrometheeIV}
\title{RPrometheeIV}
\usage{
RPrometheeIV(RPrometheeArguments)
}
\arguments{
\item{RPrometheeArguments}{An object with all RPromethee arguments. It's
important that \code{parms} argument isn't compound of NA values. See
\code{\link{RPrometheeConstructor}} for more information.}
}
\value{
\itemize{
  \item{PhiPlus} {The resulting PhiPlus from the alternatives for all
  criterias.}
  \item{PhiMinus} {The resulting PhiMinus from the alternatives for all
  criterias}
  \item{Index} {The index resulting from the lp solution.}
  \item{alternatives} {The alternatives names.}
  \item{criterias} {The criterias names.}
  \item{data} {The data used corresponding to criterias and alternatives.}
 }
}
\description{
Proposed by Brans and Vincke (1985), PROMETHEE II method aims to solve
  sorting problems. The PROMETHEE II method performs a total ordering of the
  alternatives set by calculating the net outranking flow (HENDRIKS et al.,
  1992), with the objective of solving the problem that no unambiguous
  solution can be given due to incomparability.
}
\references{
\itemize{
      \item
      M. Behzadian et al. \cr
      \emph{PROMETHEE: A comprehensive literature review on methodologies and
       applications}\cr
       European Journal of Operational Research v. 200, p.198-215, 2010.\cr
      \url{https://www.sciencedirect.com/science/article/abs/pii/S0377221709000071}
      \item
      J. P. Brans, Ph. Vincke\cr
      \emph{A Preference Ranking Organisation Method: (The PROMETHEE Method
      for Multiple Criteria Decision-Making)}\cr
      Management science, v. 31, n. 6, p. 647-656, 1985.\cr
      \url{https://pdfs.semanticscholar.org/edd6/f5ae9c1bfb2fdd5c9a5d66e56bdb22770460.pdf}

      \item
      J. P. Brans, B. Mareschal \cr
      \emph{PROMETHEE methods. In: Figueria J, Greco S, Ehrgott M (eds)
      Multiple criteria decision analysis: state of the art surveys.}\cr
      Springer Science, Business Media Inc., Boston pp 163–195.\cr
      \url{http://www.springer.com/la/book/9780387230818}

      \item
      Tsuen-Ho Hsu, Ling-Zhong Lin\cr
      \emph{Using Fuzzy Preference Method for Group Package Tour Based on the
      Risk Perception}.\cr
      Group Decision and Negotiation, v. 23, n. 2, p. 299-323, 2014.\cr
      \url{http://link.springer.com/article/10.1007/s10726-012-9313-7}
   }
}
\seealso{
Other RPromethee methods: \code{\link{NetworkPlot}},
  \code{\link{PrometheeIIIPlot}},
  \code{\link{PrometheeIIPlot}},
  \code{\link{PrometheeIPlot}},
  \code{\link{PrometheeIVPlot}},
  \code{\link{RPrometheeConstructor}},
  \code{\link{RPrometheeIII}}, \code{\link{RPrometheeII}},
  \code{\link{RPrometheeIVKernel}},
  \code{\link{RPrometheeI}}, \code{\link{RPrometheeV}},
  \code{\link{SensitivityAnalysis}},
  \code{\link{WalkingWeightsPlot}},
  \code{\link{plot,RPrometheeI-method}},
  \code{\link{plot,RPrometheeII-method}},
  \code{\link{plot,RPrometheeIII-method}}
}
\author{
Pedro Henrique Melo Albuquerque, \email{pedroa@unb.br}

Gustavo Monteiro Pereira, \email{monteirogustavop@gmail.com}
}
\keyword{decision-analysis}
\keyword{decision-method}
\keyword{mcda}
\keyword{promethee}
