% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrometheeS4.R
\docType{methods}
\name{show,RPrometheeArguments-method}
\alias{show,RPrometheeArguments-method}
\title{Shows a RPromethee object.}
\usage{
\S4method{show}{RPrometheeArguments}(object)
}
\arguments{
\item{object}{A RPromethee object.}
}
\description{
Shows data and some results for \code{RPrometheeArguments} object.
}
