% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrometheeS4.R
\docType{methods}
\name{show,RPrometheeI-method}
\alias{show,RPrometheeI-method}
\title{Show a RPromethee object}
\usage{
\S4method{show}{RPrometheeI}(object)
}
\arguments{
\item{object}{A RPromethee object.}
}
\description{
Shows data and some results for \code{RPrometheeI}.
}
