% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrometheeS4.R
\docType{methods}
\name{UpdateRPrometheeAlternatives,SensitivityAnalysis-method}
\alias{UpdateRPrometheeAlternatives,SensitivityAnalysis-method}
\title{UpdateRPrometheeAlternatives}
\usage{
\S4method{UpdateRPrometheeAlternatives}{SensitivityAnalysis}(object,
  alternatives)
}
\arguments{
\item{object}{An object from a RPromethee class.}

\item{alternatives}{A character vector with the alternatives new names.}
}
\description{
Updates alternatives names from RPromethee objects.
}
