% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrometheeS4.R
\docType{methods}
\name{UpdateRPrometheeArguments,RPrometheeArguments-method}
\alias{UpdateRPrometheeArguments,RPrometheeArguments-method}
\title{UpdateRPrometheeArguments}
\usage{
\S4method{UpdateRPrometheeArguments}{RPrometheeArguments}(object, element,
  newValue)
}
\arguments{
\item{object}{A \code{RPrometheeArguments} object}

\item{element}{A character value to indicate which slot is going to be
updated. The name must be exactly the same as the name of the argument.}

\item{newValue}{An object of the class of the element that is being updated.
For example, if it is \code{parms}, \code{newValue} must be a numeric vector.
A character vector with the alternatives new names.}
}
\description{
Updates slots from \code{RPrometheeArguments} objects.
}
