% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PrometheeIVKernel}
\alias{PrometheeIVKernel}
\title{Create the Kernel matrix}
\usage{
PrometheeIVKernel(datMat, vecWeights, prefFunction, parms, band, normalize)
}
\arguments{
\item{datVec}{Column of the dataset}

\item{int}{Type od preference function}
}
\value{
Preference Matrix
}
\description{
Create the Kernel matrix
}
