% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrometheeS4.R
\name{SensitivityAnalysis}
\alias{SensitivityAnalysis}
\title{SensitivityAnalysis}
\usage{
SensitivityAnalysis(RPrometheeArguments, method = "RPrometheeII")
}
\arguments{
\item{RPrometheeArguments}{An object with all RPromethee arguments. For
PROMETHEE IV, it's important that \code{parms} argument isn't compound of NA
values. See \code{\link{RPrometheeConstructor}} for more information.}

\item{method}{A character object used to choose how the SensitivityAnalysis is going to be calculated. The method can be \code{"RPrometheeII"} or \code{"RPrometheeIV"}. The standard is \code{"RPrometheeII"}}
}
\value{
\itemize{
  \item{Solution} {The solution resulting from linear programming problem.}
  \item{alternatives} {The alternatives names.}
  \item{criterias} {The criterias names.}
  \item{data} {The data used corresponding to criterias and alternatives.}
 }
}
\description{
Sensitivity Analysis is a method developed by Wolters & Mareschal (1995) to
  evaluate how \code{\link{RPrometheeII}} and \code{\link{RPrometheeIV}}
  results are sensitive to changes in weights of criterias. That is, how the
  solution to the decision problem can be affected by the distribution of
  criterias weights.
}
\references{
\itemize{

      \item
      M. Behzadian et al. \cr
      \emph{PROMETHEE: A comprehensive literature review on methodologies and
       applications}\cr
       European Journal of Operational Research v. 200, p.198-215, 2010.\cr
      \url{https://www.sciencedirect.com/science/article/abs/pii/S0377221709000071}
      \item
      J. P. Brans, Ph. Vincke\cr
      \emph{A Preference Ranking Organisation Method: (The PROMETHEE Method
      for Multiple Criteria Decision-Making)}\cr
      Management science, v. 31, n. 6, p. 647-656, 1985.\cr
      \url{https://pdfs.semanticscholar.org/edd6/f5ae9c1bfb2fdd5c9a5d66e56bdb22770460.pdf}

      \item
      J. P. Brans, B. Mareschal \cr
      \emph{PROMETHEE methods. In: Figueria J, Greco S, Ehrgott M (eds)
      Multiple criteria decision analysis: state of the art surveys.}\cr
      Springer Science, Business Media Inc., Boston pp 163–195.\cr
      \url{http://www.springer.com/la/book/9780387230818}

      \item
      W.T.M. Wolters, B. Mareschal\cr
      \emph{Novel types of sensitivity analysis for additive MCDM
      ethods}.\cr
      European Journal of Operational Research, v. 81, p. 281-290, 1995.\cr
      \url{https://www.sciencedirect.com/science/article/abs/pii/0377221793E0343V}
   }
}
\seealso{
Other RPromethee methods: \code{\link{NetworkPlot}},
  \code{\link{PrometheeIIIPlot}},
  \code{\link{PrometheeIIPlot}},
  \code{\link{PrometheeIPlot}},
  \code{\link{PrometheeIVPlot}},
  \code{\link{RPrometheeConstructor}},
  \code{\link{RPrometheeIII}}, \code{\link{RPrometheeII}},
  \code{\link{RPrometheeIVKernel}},
  \code{\link{RPrometheeIV}}, \code{\link{RPrometheeI}},
  \code{\link{RPrometheeV}},
  \code{\link{WalkingWeightsPlot}},
  \code{\link{plot,RPrometheeI-method}},
  \code{\link{plot,RPrometheeII-method}},
  \code{\link{plot,RPrometheeIII-method}}
}
\author{
Pedro Henrique Melo Albuquerque, \email{pedroa@unb.br}

Gustavo Monteiro Pereira, \email{monteirogustavop@gmail.com}
}
\keyword{decision-analysis}
\keyword{decision-method}
\keyword{mcda}
\keyword{promethee}
