% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrometheeS4.R
\name{UpdateRPrometheeAlternatives}
\alias{UpdateRPrometheeAlternatives}
\title{UpdateRPrometheeAlternatives}
\usage{
UpdateRPrometheeAlternatives(object, alternatives)
}
\arguments{
\item{object}{An object from a RPromethee class. It can be any of the 6
methods.}

\item{alternatives}{A character vector with the alternatives new names.}
}
\description{
Updates alternatives names from RPromethee objects.
}
\details{
It's possible to update alternatives names for: \code{RPrometheeI},
 \code{RPrometheeII}, \code{RPrometheeIII}, \code{RPrometheeIV},
 \code{RPrometheeIVKernel} and \code{RPrometheeV}
}
\seealso{
Other RPromethee methods: \code{\link{PrometheeIIIPlot}},
  \code{\link{PrometheeIIPlot}},
  \code{\link{PrometheeIPlot}},
  \code{\link{PrometheeIVPlot}},
  \code{\link{RPrometheeConstructor}},
  \code{\link{RPrometheeIII}}, \code{\link{RPrometheeII}},
  \code{\link{RPrometheeIVKernel}},
  \code{\link{RPrometheeIV}}, \code{\link{RPrometheeI}},
  \code{\link{RPrometheeV}},
  \code{\link{SensitivityAnalysis}},
  \code{\link{UpdateRPrometheeArguments}},
  \code{\link{WalkingWeightsPlot}},
  \code{\link{plot,RPrometheeI-method}}
}
\author{
Pedro Henrique Melo Albuquerque, \email{pedroa@unb.br}

Gustavo Monteiro Pereira, \email{monteirogustavop@gmail.com}
}
\keyword{decision-analysis}
\keyword{decision-method}
\keyword{mcda}
\keyword{promethee}
