% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrometheeS4.R
\name{UpdateRPrometheeArguments}
\alias{UpdateRPrometheeArguments}
\title{UpdateRPrometheeArguments}
\usage{
UpdateRPrometheeArguments(object, element, newValue)
}
\arguments{
\item{object}{A \code{RPrometheeArguments} object}

\item{element}{A character value to indicate which slot is going to be
updated. The name must be exactly the same as the name of the argument.}

\item{newValue}{An object of the class of the element that is being updated.
For example, if it is \code{parms}, \code{newValue} must be a numeric vector.}
}
\description{
Updates slots from \code{RPrometheeArguments} objects.
}
\details{
The updated arguments can be \code{datMat}, \code{vecWeights},
\code{vecMaximiz}, \code{prefFunction}, \code{parms}, \code{normalize},
\code{alphaVector}, \code{band}, \code{constraintDir} or \code{bounds}.
}
\seealso{
Other RPromethee methods: \code{\link{NetworkPlot}},
  \code{\link{PrometheeIIIPlot}},
  \code{\link{PrometheeIIPlot}},
  \code{\link{PrometheeIPlot}},
  \code{\link{PrometheeIVPlot}},
  \code{\link{RPrometheeConstructor}},
  \code{\link{RPrometheeIII}}, \code{\link{RPrometheeII}},
  \code{\link{RPrometheeIVKernel}},
  \code{\link{RPrometheeIV}}, \code{\link{RPrometheeI}},
  \code{\link{RPrometheeV}},
  \code{\link{SensitivityAnalysis}},
  \code{\link{UpdateRPrometheeAlternatives}},
  \code{\link{WalkingWeightsPlot}}
}
\author{
Pedro Henrique Melo Albuquerque, \email{pedroa@unb.br}

Gustavo Monteiro Pereira, \email{monteirogustavop@gmail.com}
}
\keyword{decision-analysis}
\keyword{decision-method}
\keyword{mcda}
\keyword{promethee}
